chore: move analytics out of root navigator
