+++ app/screens/authentication-screen/authentication-check-screen.tsx
@@ -12,8 +12,8 @@ import type { RootStackParamList } from "../../navigation/stack-param-lists"
 import { StackNavigationProp } from "@react-navigation/stack"
 
 import BitcoinBeachLogo from "../get-started-screen/bitcoin-beach-logo.png"
-import { useAuthenticationContext } from "@app/store/authentication-context"
 import { useIsAuthed } from "@app/graphql/is-authed-context"
+import { useAuthenticationContext } from "@app/navigation/navigation-container-wrapper"
 
 const styles = EStyleSheet.create({
   Logo: {
