+++ app/screens/authentication-screen/pin-screen.tsx
@@ -14,8 +14,8 @@ import { RootStackParamList } from "../../navigation/stack-param-lists"
 import { StackNavigationProp } from "@react-navigation/stack"
 import { RouteProp } from "@react-navigation/native"
 import useLogout from "../../hooks/use-logout"
-import { useAuthenticationContext } from "@app/store/authentication-context"
 import { useI18nContext } from "@app/i18n/i18n-react"
+import { useAuthenticationContext } from "@app/navigation/navigation-container-wrapper"
 
 const styles = EStyleSheet.create({
   bottomSpacer: {
