+++ app/screens/settings-screen/settings-screen.story.tsx
@@ -19,14 +19,14 @@ const propsDefault = {
 storiesOf("Settings screen", module)
   .addDecorator(withKnobs)
   .addDecorator((fn) => <StoryScreen>{fn()}</StoryScreen>)
-  .add("Not logged in", () => <SettingsScreenJSX hasToken={false} phone={undefined} username={undefined} {...propsDefault} />)
+  .add("Not logged in", () => <SettingsScreenJSX isAuthed={false} phone={undefined} username={undefined} {...propsDefault} />)
   .add("Logged in but not username", () => (
-    <SettingsScreenJSX phone="+16505551234" hasToken={true} username={undefined} {...propsDefault} />
+    <SettingsScreenJSX phone="+16505551234" isAuthed={true} username={undefined} {...propsDefault} />
   ))
   .add("Logged in with username", () => (
     <SettingsScreenJSX
       {...propsDefault}
-      hasToken={true}
+      isAuthed={true}
       phone="+16505551234"
       username="Joe"
     />
